# docs/contracts/account-recovery.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SCDC account verification and recovery",
    "version": "1.0.0",
    "description": "4 thao tác verify/resend/forgot/reset đã triển khai ngày 2026-10-05. Swagger từ source là runtime contract. Cooldown 60 giây trả accepted công khai; limiter 429 bổ sung vẫn hoãn DEC-089. Gmail SMTP worker đã có, chưa kiểm chứng giao email thật."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
    "/auth/verify-email": {
      "post": {
        "operationId": "verifyEmail",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No content"
          },
          "400": {
            "description": "Validation hoặc token sai purpose/hết hạn/đã dùng/bị thay thế; ProblemDetails theo source.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "x-scdc-implementation": "implemented",
        "x-scdc-side-effects": "Only after valid checks and transaction commit. No automatic retry of token consume."
      }
    },
    "/auth/resend-verification": {
      "post": {
        "operationId": "resendEmailVerification",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Yêu cầu được tiếp nhận; không xác nhận email tồn tại/được giao.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailRequestAccepted"
                },
                "example": {
                  "accepted": true
                }
              }
            }
          },
          "400": {
            "description": "Validation hoặc token sai purpose/hết hạn/đã dùng/bị thay thế; ProblemDetails theo source.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "x-scdc-implementation": "implemented",
        "x-scdc-side-effects": "Only after valid checks and transaction commit. No automatic retry of token consume."
      }
    },
    "/auth/forgot-password": {
      "post": {
        "operationId": "requestPasswordReset",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Yêu cầu được tiếp nhận; không xác nhận email tồn tại/được giao.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailRequestAccepted"
                },
                "example": {
                  "accepted": true
                }
              }
            }
          },
          "400": {
            "description": "Validation hoặc token sai purpose/hết hạn/đã dùng/bị thay thế; ProblemDetails theo source.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "x-scdc-implementation": "implemented",
        "x-scdc-side-effects": "Only after valid checks and transaction commit. No automatic retry of token consume."
      }
    },
    "/auth/reset-password": {
      "post": {
        "operationId": "resetPassword",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No content"
          },
          "400": {
            "description": "Validation hoặc token sai purpose/hết hạn/đã dùng/bị thay thế; ProblemDetails theo source.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "x-scdc-implementation": "implemented",
        "x-scdc-side-effects": "Only after valid checks and transaction commit. No automatic retry of token consume."
      }
    }
  },
  "components": {
    "schemas": {
      "EmailRequest": {
        "type": "object",
        "required": [
          "email"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "x-scdc-utf16-length": {
              "min": 1,
              "max": 254
            },
            "description": "Trim rồi chuyển chữ thường theo DEC-063; không bỏ dấu chấm/+tag."
          }
        },
        "additionalProperties": true
      },
      "VerifyEmailRequest": {
        "type": "object",
        "required": [
          "token"
        ],
        "properties": {
          "token": {
            "type": "string",
            "writeOnly": true,
            "minLength": 20,
            "maxLength": 512,
            "description": "Token opaque ASCII; không ghi log/cache hoặc gửi qua query API."
          }
        },
        "additionalProperties": true
      },
      "ResetPasswordRequest": {
        "type": "object",
        "required": [
          "token",
          "newPassword"
        ],
        "properties": {
          "token": {
            "type": "string",
            "writeOnly": true,
            "minLength": 20,
            "maxLength": 512,
            "description": "Token opaque ASCII; không ghi log/cache hoặc gửi qua query API."
          },
          "newPassword": {
            "type": "string",
            "writeOnly": true,
            "x-scdc-utf16-length": {
              "min": 8,
              "max": 128
            },
            "x-scdc-password-rule": "At least one char.IsLetter and one char.IsDigit; do not trim."
          }
        },
        "additionalProperties": true
      },
      "EmailRequestAccepted": {
        "type": "object",
        "required": [
          "accepted"
        ],
        "properties": {
          "accepted": {
            "const": true
          },
          "developmentResetToken": {
            "type": "null",
            "description": "Chỉ null hoặc bỏ trường trong production; token Development nằm ở runtime contract riêng."
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "errorCode",
          "traceId"
        ],
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "detail": {
            "type": [
              "string",
              "null"
            ]
          },
          "instance": {
            "type": [
              "string",
              "null"
            ]
          },
          "errorCode": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        },
        "additionalProperties": true
      }
    }
  },
  "x-scdc-status": "implemented-core",
  "x-scdc-decisions": [
    "DEC-051",
    "DEC-063",
    "DEC-064",
    "DEC-065",
    "DEC-067",
    "DEC-089",
    "DEC-113",
    "DEC-114"
  ]
}